got rid of art component combined with art view
