+++ src/views/HomeView.vue
@@ -10,12 +10,11 @@
 
 <script>
 // @ is an alias to /src
-import Art from '@/components/art.vue'
 
 export default {
   name: 'HomeView',
   components: {
-    Art
+    
   }
 }
 </script>
